feat(CarName): 자동차 이름 문자열 포장
